Extract Label component into own JS module
Allow to reuse the same Label component in other charts.
